pyproject.toml
[project]
name = "boilerplate-python-component"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksei Maksimov",email = "aleksei.maksimov@axenix.pro"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pre-commit = "^4.2.0"
